Backend/requirements.txt
# AI PPT Generator - Backend
# FastAPI + PostgreSQL + Redis + Celery

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
slowapi==0.1.9
email-validator==2.3.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Cache
redis==5.0.1

# Task Queue
celery==5.3.6

# AI Clients
openai==1.12.0
anthropic==0.18.0
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
cryptography==42.0.2

# Export
python-pptx==0.6.23

# Utils
aiofiles==23.2.1
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
